Move authenticated part into its own router
