+++ src/app/(authenticated)/(globalModals)/layout.tsx
