+++ src/app/(authenticated)/(globalModals)/settings/layout.tsx
