+++ src/app/(authenticated)/(globalModals)/talks/[slug]/_components/TalkAttribute/index.tsx
