+++ src/app/(authenticated)/(globalModals)/talks/add/_components/SearchItem/index.tsx
