+++ src/app/(authenticated)/(globalModals)/talks/scan/page.tsx
